cubfits-package {cubfits}	R Documentation

Codon Bias Usage Fits

Description

Estimating mutation and selection coefficients on synonymous codon bias usage based on models of ribosome overhead cost (ROC). Multinomial logistic regression and Markov Chain Monte Carlo are used to estimate and predict protein production rates with/without the presence of expressions and measurement errors.

Details

Package: 	 cubfits

Type: 	 Package

License: 	 Mozilla Public License 2.0

LazyLoad: 	 yes

The install command is simply as

> R CMD INSTALL cubfits_*.tar.gz

from a command mode or

R> install.packages("cubfits")

inside an R session.

Author(s)

Wei-Chen Chen wccsnow@gmail.com, Russell Zaretzki, William Howell, Drew Schmidt, and Michael Gilchrist.

References

https://github.com/snoweye/cubfits/

See Also

init.function(), cubfits(), cubpred(), and cubappr().

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))

demo(roc.train, 'cubfits', ask = F, echo = F)
demo(roc.pred, 'cubfits', ask = F, echo = F)
demo(roc.appr, 'cubfits', ask = F, echo = F)

## End(Not run)

[Package cubfits version 0.1-2 Index]