CUB Model Approximation {cubfits}	R Documentation

Codon Usage Bias Approximation for ORFs without Expression

Description

This function provides codon usage bias approximation with observed ORFs but without any expressions.

Usage


  cubappr(reu13.df.obs, phi.pred.Init, y, n,
          nIter = 1000,
          b.Init = NULL, init.b.Scale = .CF.CONF$init.b.Scale,
              b.DrawScale = .CF.CONF$b.DrawScale,
              b.RInit = NULL,
          p.Init = NULL, p.nclass = .CF.CONF$p.nclass,
              p.DrawScale = .CF.CONF$p.DrawScale,
          phi.pred.DrawScale = .CF.CONF$phi.pred.DrawScale,
          model = .CF.CT$model[1], model.Phi = .CF.CT$model.Phi[1],
          adaptive = .CF.CT$adaptive[1],
          verbose = .CF.DP$verbose,
          iterThin = .CF.DP$iterThin, report = .CF.DP$report)

Arguments

reu13.df.obs
a reu13.df object, ORFs information.

phi.pred.Init
a phi.Obs object, temporarily initial of expression without measurement errors.

y
a y object, codon counts.

n
a n object, total codon counts.

nIter
number of iterations after burn-in iterations.

b.Init
initial values for parameters b.

init.b.Scale
for initial b if b.Init = NULL.

b.DrawScale
scaling factor for adaptive MCMC with random walks when drawing new b.

b.RInit
initial values (in a list) for R matrices of parameters b yielding from QR decomposition of vglm() for the variance-covariance matrix of b.

p.Init
initial values for hyper-parameters.

p.nclass
number of components for model.Phi = "logmixture".

p.DrawScale
scaling factor for adaptive MCMC with random walks when drawing new sigma.Phi.

phi.pred.DrawScale
scaling factor for adaptive MCMC with random walks when drawing new Phi of predicted set.

model
model to be fitted, currently "roc" only.

model.Phi
prior model for Phi, currently "lognormal".

adaptive
adaptive method of MCMC for proposing new b and Phi.

verbose
print iteration messages.

iterThin
thinning iterations.

report
number of iterations to report more information.

Details

Total number of MCMC iterations is nIter + 1, but the outputs may be thinned to nIter / iterThin + 1 iterations.

Temporary result dumping may be controlled by .CF.DP.

Value

A list contains three big lists of MCMC traces including: b.Mat for mutation and selection coefficients of b, p.Mat for hyper-parameters, and phi.Mat for expected expression values Phi. All lists are of length nIter / iterThin + 1 and each element contains the output of each iteration.

All lists also can be binded as trace matrices, such as via do.call("rbind", b.Mat) yielding a matrix of dimension number of iterations by number of parameters. Then, those traces can be analyzed further via other MCMC packages such as coda.

Note

Note that phi.pred.Init need to be normalized to mean 1.

p.DrawScale may cause scaling prior if adaptive MCMC is used, and it can result in non-exits of equilibrium distribution.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

https://github.com/snoweye/cubfits/

See Also

DataIO, DataConverting, cubfits() and cubpred().

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))

demo(roc.appr, 'cubfits', ask = F, echo = F)

## End(Not run)

[Package cubfits version 0.1-2 Index]