Plotprxy {cubfits}	R Documentation

Predictive X-Y Plot

Description

This utility function provides a basic plot of production rates.

Usage


  plotprxy(x, y, x.ci = NULL, y.ci = NULL,
           log10.x = TRUE, log10.y = TRUE,
           add.lm = TRUE, add.one.to.one = TRUE, weights = NULL,
           add.legend = TRUE,
           xlim = NULL, ylim = NULL,
           xlab = "Predicted Production Rate (log10)",
           ylab = "Observed Production Rate (log10)",
           main = NULL)

Arguments

x
expression values.

y
expression values, of the same length of x.

x.ci
confidence interval of x, of dimension length{x} * 2, for outliers labeling.

y.ci
confidence interval of y, of dimension length{y} * 2, for outliers labeling.

log10.x
log10() and mean transformation of x axis.

log10.y
log10() and mean transformation of y axis.

add.lm
if add lm() fit.

add.one.to.one
if add one-to-one line.

weights
weights to lm().

add.legend
if add default legend.

xlim
limits of x-axis.

ylim
limits of y-axis.

xlab
an option passed to plot().

ylab
an option passed to plot().

main
an option passed to plot().

Details

As the usual X-Y plot where x and y are expression values.

If add.lm = TRUE and weights are given, then both ordinary and weighted least squares results will be plotted.

Value

A scatter plot with a fitted lm() line and R squared value.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

https://github.com/snoweye/cubfits/

See Also

plotbin() and plotmodel().

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))

y.scuo <- convert.y.to.scuo(ex.train$y)
SCUO <- calc_scuo_values(y.scuo)$SCUO
plotprxy(ex.train$phi.Obs, SCUO)

## End(Not run)

[Package cubfits version 0.1-2 Index]