Plotbin {cubfits}	R Documentation

Plot Binning Results

Description

Plot binning results to visualize the effects of mutation and selection along with expression levels empirically.

Usage


  prop.bin.roc(reu13.df, phi.Obs = NULL, nclass = 20, bin.class = NULL,
               weightedCenters = TRUE, logBins = FALSE)

  plotbin(ret.bin, ret.model = NULL, main = NULL,
          xlab = "Production Rate (log10)", ylab = "Proportion",
          xlim = NULL, lty = 1, x.log10 = TRUE, stderr = FALSE, ...)

Arguments

reu13.df
a reu13.df object.

phi.Obs
a phi.Obs object.

nclass
number of binning classes across the range of phi.Obs.

bin.class
binning proportion, e.g. c(0, seq(0.05, 0.95, length = nclass), 1).

ret.bin
binning results from prop.bin.roc().

weightedCenters
if centers are weighted.

logBins
if use log scale for bin.

ret.model
model results from prop.model.roc().

main
an option passed to plot().

xlab
an option passed to plot().

ylab
an option passed to plot().

xlim
range of X-axis.

lty
line type if ret.model is provided.

x.log10
log10() transformation of X-axis.

stderr
plot stand error instead of stand deviation.

...
options passed to plot().

Details

The function plotbin() plots the binning results ret.bin returned from prop.bin.roc(). Fitted curves may be added if ret.model is provided which can be obtained from prop.model.roc().

plotaddmodel() can append model later if ret.model is not provided to plotbin().

Currently, only ROC model is supported. Colors are controlled by .CF.PT.

Value

A binning plot is drawn.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

https://github.com/snoweye/cubfits/

See Also

plotmodel() and prop.model.roc().

Examples


## Not run: 
demo(plotbin, 'cubfits', ask = F, echo = F)

## End(Not run)

[Package cubfits version 0.1-2 Index]