Plotmodel {cubfits}	R Documentation

Plot Fitted Models

Description

Plot model results to visualize the effects of mutation and selection along with expression levels. The model can be fitted by MCMC or multinomial logistic regression.

Usage


  prop.model.roc(b.Init, phi.Obs.lim = c(0.01, 10), phi.Obs.scale = 1,
                 nclass = 40, x.log10 = TRUE)

  plotmodel(ret.model, main = NULL,
            xlab = "Production Rate (log10)", ylab = "Proportion",
            xlim = NULL, lty = 1, x.log10 = TRUE, ...)

  plotaddmodel(ret.model, lty, u.codon = NULL, color = NULL,
               x.log10 = TRUE)

Arguments

b.Init
a b object.

phi.Obs.lim
range of phi.Obs.

phi.Obs.scale
optional scaling factor.

nclass
number of binning classes across the range of phi.Obs.

x.log10
log10() transformation of X-axis.

ret.model
model results from prop.model.roc().

main
an option passed to plot().

xlab
an option passed to plot().

ylab
an option passed to plot().

xlim
range of X-axis.

lty
line type.

u.codon
unique synonymous codon names.

color
a color vector for unique codon, typically returns of the internal function get.color().

...
options passed to plot().

Details

The function plotmodel() plots the fitted curves obtained from prop.model.roc().

The function plotaddmodel() can append model curves to a binning plot provided unique synonymous codons and colors are given. This function is nearly for an internal call within plotmodel(), but is exported and useful for workflow.

Currently, only ROC model is supported. Colors are controlled by .CF.PT.

Value

A fitted curve plot is drawn.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

https://github.com/snoweye/cubfits/

See Also

plotbin(), prop.bin.roc(), and prop.model.roc().

Examples


## Not run: 
demo(plotbin, 'cubfits', ask = F, echo = F)

## End(Not run)

[Package cubfits version 0.1-2 Index]