Cedric Convergence Utilities {cubfits}	R Documentation

Cedric Convergence Utilities

Description

This utility function provides convergence related functions by Cedric.

Usage


  cubmultichain(cubmethod, reset.qr, seeds=NULL,
                teston=c("phi", "sphi"), swap=0, swapAt=0.05, monitor=NULL, 
                min=0, max=160000, nchains=2, conv.thin=10,
                eps=0.1, ncores=2, ...)
                          
  cubsinglechain(cubmethod, frac1=0.1, frac2=0.5, reset.qr,
                 seed=NULL, teston=c("phi", "sphi"), monitor=NULL, 
                 min=0, max=160000, conv.thin=10, eps=1, ...)

Arguments

cubmethod
String to choose method. Options are "cubfits", "cubappr", "cubpred"

reset.qr
recalculate QR decomposition matrix of covariance matrix until reset.qr samples are reached

swap
proportion of b matrix parameters to be swaped between convergence checks

swapAt
difference (L1-norm) between two consequtive convergence test leading to a swap in the b matrix

seeds
Vector of seed for random number generation

seed
Seed for random number generation

teston
Select data to test convergence on

monitor
A function to monitor the progress of the MCMC. The fucntions expects the result object and for cubmultichain an index i. (cubmultichain call: monitor(x,i), cubsinglechain call: monitor(x))

min
Minimum samples to be obtained. eps is ignored until number of samples reaches min

max
Maximum samples to be obtained. eps is ignored after max samples is obtained

eps
Convergence criterium

conv.thin
thinning of samples before performing convergence test

nchains
number of chains to run in parallel

ncores
number of cores to use for parallel execution of chains

frac1
fraction of samples at the beginning of set for Geweke test

frac2
fraction of samples at the end of set for Geweke test

...
named arguments for functions "cubfits", "cubappr" or "cubpred"

Details

under development

Value

under development

Author(s)

Cedric Landerer cedric.landerer@gmail.com.

References

https://github.com/clandere/cubfits/

See Also

cubfits, cubappr, cubpred

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))


## End(Not run)

[Package cubfits version 0.1-2 Index]