SCUO Index {cubfits}	R Documentation

Function for Synonymous Codon Usage Order (SCUO) Index

Description

Calculate the Synonymous Codon Usage Order (SCUO) index for each gene. Used as a substitute for expression in cases of without expression measurements.

Usage


  calc_scuo_values(codon.counts)

Arguments

codon.counts
an object of format scuo.

Details

This function computes SCUO index for each gene. Typically, this method is completely based on entropy and information theory to estimate expression values of sequences according to their codon information.

Value

SCUO indices are returned.

Author(s)

Drew Schmidt.

References

http://www.tandfonline.com/doi/abs/10.1080/03081070500502967

Wan X.-F., Zhou J., Xu D. “CodonO: a new informatics method for measuring synonymous codon usage bias within and across genomes” International Journal of General Systems Vol. 35, Iss. 1, 2006.

See Also

scuo.random(), calc_cai_values(), calc_scu_values().

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))

y.scuo <- convert.y.to.scuo(ex.train$y)
SCUO <- calc_scuo_values(y.scuo)$SCUO
plotprxy(ex.train$phi.Obs, SCUO, ylab = "SCUO (log10)")

## End(Not run)

[Package cubfits version 0.1-2 Index]