Datasets {cubfits}	R Documentation

Datasets for Demonstrations

Description

Examples of toy data to test and demonstrate cubfits.

Usage


  b.Init
  ex.test
  ex.train

Format

All are in list formats.

Details

b.Init contains two sets (roc and rocnse) of initial coefficients including mutation and selection parameters for 3 amino acids 'A', 'C', and 'D' in matrix format. Both sets are in b format.

ex.train contains a training set of 100 sequences including 3 reu13.df (codon counts in reu13 data frame format divided by amino acids), 3 y (codon counts in simplified data frame format divided by amino acids), 3 n (total amino acid counts in vector format divided by amino acids), and phi.Obs (observed phi values in vector format).

ex.test contains a testing set of the other 100 sequences in the same format of ex.train.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

https://github.com/snoweye/cubfits/

See Also

init.function(), cubfits(), cubpred(), and cubappr().

Examples


## Not run: 
suppressMessages(library(cubfits, quietly = TRUE))

str(b.Init)
str(ex.test)
str(ex.train)

## End(Not run)

[Package cubfits version 0.1-2 Index]